+++ src/pools/cluster/fixed.ts
@@ -107,7 +107,7 @@ export class FixedClusterPool<
 
   /** @inheritDoc */
   protected get full (): boolean {
-    return this.workerNodes.length === this.numberOfWorkers
+    return this.workerNodes.length >= this.numberOfWorkers
   }
 
   /** @inheritDoc */