+++ src/pools/thread/dynamic.ts
@@ -42,7 +42,7 @@ export class DynamicThreadPool<
 
   /** @inheritDoc */
   protected get full (): boolean {
-    return this.workerNodes.length === this.max
+    return this.workerNodes.length >= this.max
   }
 
   /** @inheritDoc */